+++ openTelemetry/CHANGELOG.md
@@ -6,8 +6,9 @@ This file documents changes to the `workbench-openTelemetry` library, including 
 
 ### Added
 - Add `OpenTelemetryMetrics`
+- Add `registerTracing`
 
 ### Changed
 - Bump `scalatest` to `3.1.1`
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-opentelemetry" % "0.1-73d6a64"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-opentelemetry" % "0.1-TRAVIS-REPLACE-ME"`